requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.8.1
